requirements.txt
Django==3.2.23
python-dotenv
requests
geopy